# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "The transactional core: agreements, orders, the shipment lifecycle and invoices.\\n\\nStatus changes go through state machines rather than accepting an arbitrary status string, so an illegal transition is refused with 409 rather than written.",
        "title": "Karlo Business API",
        "termsOfService": "https://karlo.co.id/terms",
        "contact": {
            "name": "Karlo Engineering",
            "email": "engineering@karlo.co.id"
        },
        "version": "1.0"
    },
    "host": "localhost:5003",
    "basePath": "/api/v1",
    "paths": {
...
        "/orders/{orderId}/shipment": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Shipments"
                ],
                "summary": "Get an order's shipment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "orderId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_karlo_business-service_internal_models.Shipment"
                        }
                    }
                }
            }
        },
        "/shipments/{id}/documents": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Shipments"
                ],
                "summary": "List shipment documents",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shipment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Shipments"
                ],
                "summary": "Attach a shipment document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shipment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/shipments/{id}/status": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Shipments"
                ],
                "summary": "Advance shipment status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shipment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_karlo_business-service_internal_models.Shipment"
                        }
                    }
                }
            }
        },
        "/shipments/{id}/transitions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Shipments"
                ],
                "summary": "Available shipment transitions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shipment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_karlo_business-service_internal_models.Shipment": {
            "type": "object",
            "properties": {
                "arrivedLoadingAt": {
                    "type": "string"
                },
                "arrivedUnloadingAt": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "distanceMeters": {
                    "type": "integer"
                },
                "driverUserId": {
                    "type": "string"
                },
                "finishedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "loadingFinishedAt": {
                    "type": "string"
                },
                "loadingLatitude": {
                    "type": "number"
                },
                "loadingLongitude": {
                    "type": "number"
                },
                "loadingStartedAt": {
                    "type": "string"
                },
                "loadingWithinGeofence": {
                    "type": "boolean"
                },
                "orderId": {
                    "type": "string"
                },
                "startedToLoadingAt": {
                    "type": "string"
                },
                "startedToUnloadingAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "statusAlias": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "string"
                },
                "tollCost": {
                    "type": "number"
                },
                "truckId": {
                    "type": "string"
                },
                "unloadingFinishedAt": {
                    "type": "string"
                },
                "unloadingLatitude": {
                    "type": "number"
                },
                "unloadingLongitude": {
                    "type": "number"
                },
                "unloadingStartedAt": {
                    "type": "string"
                },
                "unloadingWithinGeofence": {
                    "type": "boolean"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "RS256 access token issued by the authentication service, as \"Bearer \u003ctoken\u003e\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}